pyproject.toml
[project]
name = "lora-judge"
version = "0.1.0"
description = "QLoRA pipeline for training and evaluating LLMs on judge tasks."
requires-python = ">=3.11"
dependencies = [
  "accelerate>=1.1.0",
  "bitsandbytes>=0.45.0",
  "datasets>=3.1.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "peft>=0.14.0",
  "pydantic>=2.10.0",
  "pydantic-ai-slim[openai]>=0.0.25",
  "pydantic-evals>=0.0.25",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0.0",
  "scikit-learn>=1.5.0",
  "torch>=2.5.0",
  "transformers>=4.47.0",
  "trl>=0.12.0"
]

[project.scripts]
judge-train = "l4_qlora_judge.train:main"
judge-infer = "l4_qlora_judge.infer_submission:main"
judge-eval = "l4_qlora_judge.eval_pydantic_evals:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/l4_qlora_judge"]

[tool.uv]
package = true
